feat: api 에러 처리 로직 분리
