+++ apps/web/src/lib/api-client.ts
@@ -1,53 +1,5 @@
-import Axios, { AxiosError, AxiosRequestConfig } from 'axios';
+import Axios, { AxiosRequestConfig } from 'axios';
 import { useAuthStore } from '@/stores/authStore';
-
-export class ApiError extends Error {
-  status?: number;
-  data?: unknown;
-  url?: string;
-
-  constructor(
-    message: string,
-    status?: number,
-    data?: unknown,
-    url?: string,
-  ) {
-    super(message);
-    this.name = 'ApiError';
-    this.status = status;
-    this.data = data;
-    this.url = url;
-    Object.setPrototypeOf(this, ApiError.prototype);
-  }
-
-  static fromAxios(error: AxiosError) {
-    return new ApiError(
-      error.message || '요청 처리 중 오류가 발생했습니다.',
-      error.response?.status,
-      error.response?.data,
-      error.config?.url,
-    );
-  }
-
-  static wrap(error: unknown): ApiError {
-    if (error instanceof ApiError) return error;
-    if (Axios.isAxiosError(error)) return ApiError.fromAxios(error);
-    if (error instanceof Error) return new ApiError(error.message);
-    return new ApiError('알 수 없는 오류가 발생했습니다.');
-  }
-
-  isAuthError() {
-    return this.status === 401;
-  }
-
-  isServerError() {
-    return (this.status ?? 0) >= 500;
-  }
-
-  isNetworkError() {
-    return this.status === undefined;
-  }
-}
 
 export const AXIOS_INSTANCE = Axios.create({
   baseURL:
@@ -89,9 +41,8 @@ AXIOS_INSTANCE.interceptors.response.use(
       console.error(`[API Error ${status}] ${url}`, data);
     }
 
-    // Handle 401 unauthorized errors
+    // 401 Unauthorized 처리
     if (error.response?.status === 401) {
-      // Access Token 만료 - 자동으로 refresh 시도는 axios.ts에서 처리
       if (typeof window !== 'undefined') {
         console.log('[API] 401 Unauthorized - 토큰 만료');
       }
